pyproject.toml
[tool.poetry]
name = "git-multi-exec"
version = "0.1.0"
description = ""
authors = ["Jonas Rosenboom <jrosenboom@checkpoint.com>"]

[tool.poetry.scripts]
git-multi-exec = "git_multi_exec.cli:cli"

[tool.poetry.dependencies]
python = "^3.10"
atlassian-python-api = "^3.32.1"
pygit2 = "^1.11.1"
octokitpy = "^0.15.0"
python-gitlab = "^3.12.0"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
bandit = "^1.7.4"
black = "^22.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
